MDL-66118 hub: Drop support for the hub's display_homepage callback
There should never be any exceptional hacks like this for any particular
plugin. The hub sites can make use of a wide range of other and cleaner
mechanisms - such as a custom theme or the customfrontpageinclude
directive - to implement a customized front page.

Plus, we've dropped support for alternative hubs anyway.