MDL-67097 libraries: fix comment parsing in php-css-parser library
This fixes a bug when parsing the following rule-level and list-level
comments.

div {/*rtl:ignore*/left:10px;/*rtl:ignore*/ display:inline;}
/*rtl:ignore*/div {left:10px;} /*rtl:ignore*/ a {left:10px}

The latter rtl directive was being swallowed by the parser in these
cases and was not being properly allocated as a property of the rule
or csslist.

This restores the correct behaviour of the rtlcss library, which depends
on php-css-parser.