MDL-69723 assign: Add # need grading to summary
For teachers viewing the assignment summary (/mod/assign/index.php),
it shows a '-' in the 'Grade' column. Instead of showing a '-', display
the number of items that need grading.

Augment by creating a link to the grading page so that teachers can see
the outstanding assignments more quickly.

Also attempted to write phpunit tests that verify values for ungraded
submissions for teachers.