MDL-70148 behat: Wait after opening the user menu
The version of Boost used in Moodle 3.5 does not have any pendingJS
checks and it is not possible to add them.

We need to use the poor-person's alternative and fall back to a short
wait to allow the menu to open before interacting with it.

This may still allow for failures on extremely slow systems but the risk
is extremely minimal.